git-svn-id: svn+ssh://combine-101/home/svn_bio/shrimp/trunk@702 450597fd-3c47-0410-84d8-f85f2794411d

+++ mergesam/merge_sam.h
@@ -25,7 +25,7 @@ struct output_filter {
 	bool strata;
 	bool unaligned;
 	bool half_paired;
-	double isize;
+	int isize;
 	bool use_isize;	
 	bool detect_isize;
 	size_t max_isize;
